+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [ENHANCEMENT] Added memory ballast. #198
 * [ENHANCEMENT] Added support for Redis cached repository. #202
 * [ENHANCEMENT] Cached VCard repository. #203
+* [ENHANCEMENT] Cached Last repository. #204
 * [CHANGE] Introduced measured repository transaction type. #200
 * [CHANGE] Use PgSQL locker. #201
 * [BUGFIX] Fix S2S db key check when nop KV is used. #199